+++ lib/browser/browser.ts
@@ -68,7 +68,8 @@ import {
   FingerprintAPI,
   UserClaims, 
   SigninWithRedirectOptions,
-  SignInWithCredentialsOptions
+  SignInWithCredentialsOptions,
+  Tokens
 } from '../types';
 import fingerprint from './fingerprint';
 import { postToTransaction } from '../tx';
@@ -137,10 +138,12 @@ class OktaAuthBrowser extends OktaAuthBase implements OktaAuth, SignoutAPI {
       redirectUri: toAbsoluteUrl(args.redirectUri, window.location.origin),
       postLogoutRedirectUri: args.postLogoutRedirectUri,
       responseMode: args.responseMode,
+      responseType: args.responseType,
       transformErrorXHR: args.transformErrorXHR,
       cookies: getCookieSettings(this, args),
       scopes: args.scopes,
-      transformAuthState: args.transformAuthState
+      transformAuthState: args.transformAuthState,
+      restoreOriginalUri: args.restoreOriginalUri
     });
   
     this.userAgent = getUserAgent(args, `okta-auth-js/${SDK_VERSION}`);
@@ -257,7 +260,7 @@ class OktaAuthBrowser extends OktaAuthBase implements OktaAuth, SignoutAPI {
     });
   }
 
-  async signInWithRedirect({ fromUri, ...additionalParams }: SigninWithRedirectOptions = {}) {
+  async signInWithRedirect({ originalUri, ...additionalParams }: SigninWithRedirectOptions = {}) {
     if(this._pending.handleLogin) { 
       // Don't trigger second round
       return;
@@ -267,8 +270,8 @@ class OktaAuthBrowser extends OktaAuthBase implements OktaAuth, SignoutAPI {
     const { scopes, responseType } = this.options;
     try {
       // Trigger default signIn redirect flow
-      if (fromUri) {
-        this.setFromUri(fromUri);
+      if (originalUri) {
+        this.setOriginalUri(originalUri);
       }
       const params = Object.assign({
         scopes: scopes || ['openid', 'email', 'profile'],
@@ -432,23 +435,80 @@ class OktaAuthBrowser extends OktaAuthBase implements OktaAuth, SignoutAPI {
     this.tokenManager.setTokens(tokens);
   }
 
-  setFromUri(fromUri?: string): void {
-    // Use current location if fromUri was not passed
-    fromUri = fromUri || window.location.href;
-    // Store fromUri
+  setOriginalUri(originalUri?: string): void {
+    // Use current location if originalUri was not passed
+    originalUri = originalUri || window.location.href;
+    // Store originalUri
     const storage = browserStorage.getSessionStorage();
-    storage.setItem(REFERRER_PATH_STORAGE_KEY, fromUri);
+    storage.setItem(REFERRER_PATH_STORAGE_KEY, originalUri);
   }
 
-  getFromUri(): string {
+  getOriginalUri(): string {
     const storage = browserStorage.getSessionStorage();
-    const fromUri = storage.getItem(REFERRER_PATH_STORAGE_KEY) || window.location.origin;
-    return fromUri;
+    const originalUri = storage.getItem(REFERRER_PATH_STORAGE_KEY) || window.location.origin;
+    return originalUri;
   }
 
-  removeFromUri(): void {
+  removeOriginalUri(): void {
     const storage = browserStorage.getSessionStorage();
     storage.removeItem(REFERRER_PATH_STORAGE_KEY);
+  }
+
+  isLoginRedirect(): boolean {
+    return isLoginRedirect(this);
+  }
+
+  async handleLoginRedirect(tokens?: Tokens): Promise<void> {
+    const handleRedirect = async ({ isPending }) => {
+      if (isPending) {
+        return;
+      }
+
+      // Unsubscribe listener
+      this.authStateManager.unsubscribe(handleRedirect);
+
+      // Get and clear originalUri from storage
+      const originalUri = this.getOriginalUri();
+      this.removeOriginalUri();
+
+      // Redirect to originalUri
+      const { restoreOriginalUri } = this.options;
+      if (restoreOriginalUri) {
+        await restoreOriginalUri(this, originalUri);
+      } else {
+        window.location.replace(originalUri);
+      }
+    };
+
+    // Handle redirect after authState is updated 
+    this.authStateManager.subscribe(handleRedirect);
+
+    // Store tokens and update AuthState by the emitted events
+    if (tokens) {
+      this.tokenManager.setTokens(tokens);
+    } else if (this.isLoginRedirect()) {
+      await this.storeTokensFromRedirect();
+    } else {
+      this.authStateManager.unsubscribe(handleRedirect);
+    }
+  }
+
+  isPKCE(): boolean {
+    return !!this.options.pkce;
+  }
+
+  hasResponseType(responseType: string): boolean {
+    let hasResponseType = false;
+    if (Array.isArray(this.options.responseType) && this.options.responseType.length) {
+      hasResponseType = this.options.responseType.indexOf(responseType) >= 0;
+    } else {
+      hasResponseType = this.options.responseType === responseType;
+    }
+    return hasResponseType;
+  }
+
+  isAuthorizationCodeFlow(): boolean {
+    return this.hasResponseType('code');
   }
 }